refactor(Deck): DeckGenreator 기능들 Deck으로 위임
