+++ src/main/java/domain/card/Deck.java
@@ -1,5 +1,6 @@
 package domain.card;
 
+import java.util.Collections;
 import java.util.List;
 import java.util.Stack;
 
@@ -7,8 +8,27 @@ public class Deck {
 
     public final Stack<Card> cards;
 
-    public Deck(final Stack<Card> cards) {
-        this.cards = cards;
+    public Deck() {
+        this.cards = generateCards();
+    }
+
+    private Stack<Card> generateCards() {
+        Stack<Card> cards = new Stack<>();
+        for (Suit suit : Suit.values()) {
+            addCardBySuit(cards, suit);
+        }
+        shuffle(cards);
+        return cards;
+    }
+
+    private void addCardBySuit(List<Card> cards, Suit suit) {
+        for (Denomination denomination : Denomination.values()) {
+            cards.add(new Card(suit, denomination));
+        }
+    }
+
+    private void shuffle(List<Card> cards) {
+        Collections.shuffle(cards);
     }
 
     public Card drawCard() {
