+++ src/test/java/domain/InitGameSetterTest.java
@@ -1,11 +1,8 @@
 package domain;
 
-import static org.junit.jupiter.api.Assertions.*;
-
 import java.util.List;
 
 import domain.card.Deck;
-import domain.card.DeckGenerator;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -15,8 +12,7 @@ class InitGameSetterTest {
     @Test
     @DisplayName("플레이어들을 생성한다.")
     void generatePlayers() {
-        DeckGenerator deckGenerator = new DeckGenerator();
-        Deck deck = deckGenerator.generate();
+        Deck deck = new Deck();
         List<String> playerNames = List.of("seongha", "dino");
 
         Assertions.assertDoesNotThrow(() -> InitGameSetter.generatePlayers(deck, playerNames));
@@ -25,8 +21,7 @@ class InitGameSetterTest {
     @Test
     @DisplayName("딜러를 생성한다.")
     void generateDealer() {
-        DeckGenerator deckGenerator = new DeckGenerator();
-        Deck deck = deckGenerator.generate();
+        Deck deck = new Deck();
 
         Assertions.assertDoesNotThrow(() -> InitGameSetter.generateDealer(deck));
     }
