+++ src/test/java/domain/card/DeckGeneratorTest.java
@@ -11,8 +11,7 @@ public class DeckGeneratorTest {
     @Test
     @DisplayName("생성한 덱은 52장이다.")
     void generateDeck() {
-        DeckGenerator deckGenerator = new DeckGenerator();
-        Deck deck = deckGenerator.generate();
+        Deck deck = new Deck();
         List<Card> cards = deck.getCards();
 
         Assertions.assertThat(cards.size()).isEqualTo(52);
