+++ src/test/java/domain/card/DeckTest.java
@@ -9,8 +9,7 @@ class DeckTest {
     @Test
     @DisplayName("덱의 맨 위에 있는 카드를 뽑아서 반환한다.")
     void getCardTopOfDeck() {
-        DeckGenerator deckGenerator = new DeckGenerator();
-        Deck deck = deckGenerator.generate();
+        Deck deck = new Deck();
         int topOfDeckCardIndex = deck.cards.size() - 1;
         Card topOfDeckCard = deck.cards.get(topOfDeckCardIndex);
